TAPAS: Add hash to the Announcement "specification"
Now TAPASAnnouncement should return:
  function announcementHash() public view returns (uint256)

Also announcementURI() can return "string", but since we are
(for now) dependant of Solidity 0.4.18, we must use "bytes32"
instead of "string".

If TAPASAnnouncement.announcementURI() wants to return a string,
it will be "empty" when emitting Announced(). This might be desired
signal to show that the client software should use "string".